requirements.txt
flask
flask-login
flask-sqlalchemy
flask-migrate
authlib
numpy
pandas
Pillow
tensorflow
joblib
fpdf2
werkzeug
python-dotenv
requests
matplotlib 